requirements.txt
timm==0.4.9
transformers==4.12.5
ruamel_yaml
opencv-python
scikit-image
matplotlib
torch==1.7.1
torchvision==0.8.2
pycocotools
pycocoevalcap